% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameco.R
\name{ameco_codes}
\alias{ameco_codes}
\title{AMECO database codes}
\usage{
ameco_codes(version = "5/2025", reset = FALSE)
}
\arguments{
\item{version}{(default 5/2025) version of AMECO (either 5/xxxx or spring xxxx for spring forecast or 11/xxxx autumn xxxx for autumn forecast)}

\item{reset}{(download even if cache is present)}
}
\value{
a tibble withn AMECO codes
}
\description{
Télécharge et renvoie le fichier des codes d'AMECO.
}
