% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_ameco.R
\name{url_ameco}
\alias{url_ameco}
\title{Renvoie l'url du vintage AMECO}
\usage{
url_ameco(version = "11025")
}
\arguments{
\item{version}{(defaut "5/2025") 5/xxxx pour la version de printemps, 11/2025 pour laversion d'automne}
}
\value{
string
}
\description{
Renvoie l'url du vintage AMECO
}
\examples{
url_ameco("5/2025")
}
