% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameco.R
\name{ameco}
\alias{ameco}
\title{AMECO database}
\usage{
ameco(version = "5/2025", reset = FALSE)
}
\arguments{
\item{version}{(default 5/2025) version of AMECO (either 5/xxxx or spring xxxx for spring forecast or 11/xxxx autumn xxxx for autumn forecast)}

\item{reset}{(download even if cache is present)}
}
\value{
a tibble withn AMECO
}
\description{
Télécharge et renvoie le fichier complet d'AMECO pour le vintage demandé. Les vintages disponibles sont ceux sur le site de la Commission Eruopéen de 2011 à 2025, avec deux prévisions par an (printemps et automne) et une préivison à l'hiver 2015.
}
