% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameco.R
\name{ameco}
\alias{ameco}
\title{AMECO database}
\usage{
ameco(version = "5/2025", reset = FALSE)
}
\arguments{
\item{version}{(default 5/2025) version of AMECO (either 5/xxxx for spring forecast or 11/xxxx for autumn forecast)}

\item{reset}{(download even if cache is present)}
}
\value{
a tibble withn AMECO
}
\description{
AMECO database
}
\examples{
ameco()
}
